:recycle: Refactor the code and add new tests
