+++ api/app/Repositories/DeveloperRepository.php
@@ -18,7 +18,7 @@ class DeveloperRepository
         return $this->developer->all();
     }
 
-    public function getDeveloperById(int $id)
+    public function getDeveloperById(int $id): Developer
     {
         return $this->developer->findOrFail($id);
     }
@@ -28,11 +28,6 @@ class DeveloperRepository
         return $this->developer->create($data);
     }
 
-    public function find(int $id)
-    {
-        return $this->developer->findOrFail($id);
-    }
-
     public function update(int $id, array $data)
     {
         $developer = $this->developer->findOrFail($id);
@@ -40,13 +35,10 @@ class DeveloperRepository
         return $developer;
     }
 
-    /**
-     * @param int $id
-     * @return mixed
-     */
     public function delete(int $id)
     {
-        $developer = $this->developer->findOrFail($id);
+        $developer = $this->getDeveloperById($id);
+
         return $developer->delete();
     }
 }
